+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+UNRELEASED
+----------
+
+- Add search/filter
+
+
 1.0.0 (2026-02-17)
 ------------------